accuracy/requirements.txt
asgiref==3.9.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.5
djangorestframework==3.16.1
dtaidistance==2.3.13
fonttools==4.59.0
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.15.3
six==1.17.0
spicy==0.16.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2